[feat] modules, index.js is now cleaner
