+++ index.js
@@ -1,53 +1,8 @@
-const moment = require("moment")
-require('dotenv').config()
-const creds = require(process.env.JSON_ENV_PATH)
-const { GoogleSpreadsheet } = require('google-spreadsheet');
-const doc = new GoogleSpreadsheet(process.env.DOC_ID);
+const getSpreadSheet = require("./lib/getSpreadSheet");
+const generateReport = require("./lib/generateReport");
 (async () => {
-    await doc.useServiceAccountAuth(creds);
-    await doc.loadInfo()
-    const sheet = doc.sheetsByIndex[0];
-    await sheet.loadCells()
+    let sheet = await getSpreadSheet.getSpreadSheet();
 
-    var prevMonday = new Date();
-    prevMonday.setDate(prevMonday.getDate() - (prevMonday.getDay() + 6) % 7)
-    var formattedMonday = prevMonday.toLocaleDateString('fr-FR') // last monday
-
-    var friday = new Date();
-    friday.setDate(prevMonday.getDate() + 4)
-    var formattedFriday = friday.toLocaleDateString('fr-FR') // friday
-
-    console.log(`# Journal de travail du ${formattedMonday} au ${formattedFriday}`)
-
-    for (let i = 2; i < 1000; i++) {
-        var cell = sheet.getCellByA1(`C${i}`)
-        if(cell.formattedValue == formattedMonday) {
-            const row = await sheet.getRows()
-            var mondayRow = cell.a1Row - 2
-            for (let i = mondayRow; i < mondayRow + 5; i++) {
-                var day = row[i]
-                var date = day["Date"]
-                var frenchDay = moment(date, 'DD/MM/YYYY').locale("fr").format("dddd").charAt(0).toUpperCase() + moment(date, 'DD/MM/YYYY').locale("fr").format("dddd").slice(1)
-                console.log(`
-### ${frenchDay} ${date}
-##### Heure d'arrivée : <u>${day["Heure d’arrivée"]}</u>
-##### Heure de départ (manger) : <u>${day["Heure de départ (manger)"]}</u>
-##### Heure de retour : <u>${day["Heure de retour"]}</u>
-##### Heure de départ <u>${day["Heure de départ"]}</u>
-##### Total : <u>${day["Total"]}</u>
-#### Description
-${day["Description"]}
-`)
-                if(day["Date"] == formattedFriday) { 
-                    console.log(`
-## Total semaine : <u>${day["Total Semaine"]}</u>
-## Heures supplémentaires : <u>${day["Diff semaine"]}</u>
-                    `)
-                }
-            }
-        }
-    }
-
-    console.log(`<i><small>Generated by [jdt-generator](https://github.com/Azecko/jdt-generator) by [Azecko](https://github.com/Azecko)</small></i>`)
+    await generateReport(sheet);
     
 })();
