+++ lib/generateReport.js
@@ -1,0 +1,43 @@
+const moment = require("moment")
+module.exports = async function generateReport(sheet) {
+    var prevMonday = new Date();
+    prevMonday.setDate(prevMonday.getDate() - (prevMonday.getDay() + 6) % 7)
+    var formattedMonday = prevMonday.toLocaleDateString('fr-FR') // last monday
+
+    var friday = new Date();
+    friday.setDate(prevMonday.getDate() + 4)
+    var formattedFriday = friday.toLocaleDateString('fr-FR') // friday
+
+    console.log(`# Journal de travail du ${formattedMonday} au ${formattedFriday}`)
+    for (let i = 2; i < 1000; i++) {
+        var cell = sheet.getCellByA1(`C${i}`)
+        if(cell.formattedValue == formattedMonday) {
+            const row = await sheet.getRows()
+            var mondayRow = cell.a1Row - 2
+            for (let i = mondayRow; i < mondayRow + 5; i++) {
+                var day = row[i]
+                console.log(day)
+                var date = day["Date"]
+                var frenchDay = moment(date, 'DD/MM/YYYY').locale("fr").format("dddd").charAt(0).toUpperCase() + moment(date, 'DD/MM/YYYY').locale("fr").format("dddd").slice(1)
+                console.log(`
+    ### ${frenchDay} ${date}
+    ##### Heure d'arrivée : <u>${day["Heure d’arrivée"]}</u>
+    ##### Heure de départ (manger) : <u>${day["Heure de départ (manger)"]}</u>
+    ##### Heure de retour : <u>${day["Heure de retour"]}</u>
+    ##### Heure de départ <u>${day["Heure de départ"]}</u>
+    ##### Total : <u>${day["Total"]}</u>
+    #### Description
+    ${day["Description"]}
+    `)
+                if(day["Date"] == formattedFriday) { 
+                    console.log(`
+    ## Total semaine : <u>${day["Total Semaine"]}</u>
+    ## Heures supplémentaires : <u>${day["Diff semaine"]}</u>
+                    `)
+                }
+            }
+        }
+    }
+
+    console.log(`<i><small>Generated by [jdt-generator](https://github.com/Azecko/jdt-generator) by [Azecko](https://github.com/Azecko)</small></i>`)
+}
